requirements.txt
fastapi
uvicorn
openai
httpx
python-dotenv
requests
beautifulsoup4
